requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord.py==1.6.0
fauna-easy==0.0.1
faunadb==4.0.1
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
iso8601==0.1.14
multidict==5.1.0
pydantic==1.8.1
python-dotenv==0.15.0
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
